+++ sketch_oxide/src/sampling/mod.rs
@@ -22,11 +22,13 @@
 pub mod l0_sampler;
 pub mod priority_sampling;
 pub mod reservoir;
+pub mod reservoir_l;
 pub mod varopt;
 pub mod weighted_reservoir;
 
 pub use l0_sampler::L0Sampler;
 pub use priority_sampling::PrioritySampling;
 pub use reservoir::ReservoirSampling;
+pub use reservoir_l::ReservoirSamplingL;
 pub use varopt::VarOptSampling;
 pub use weighted_reservoir::WeightedReservoirSampling;